backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

python-jose==3.3.0
python-dotenv==1.0.1

scikit-learn==1.4.2
joblib==1.3.2
pandas==2.1.4
numpy==1.26.4

redis==5.0.4

prometheus-fastapi-instrumentator==6.1.0

# Additional dependencies for Render deployment
gunicorn==21.2.0